Clarify Nominal/Topological distinction
