+++ single-line-diagram/single-line-diagram-core/src/main/java/com/powsybl/sld/util/TopologicalStyleProvider.java
@@ -150,8 +150,8 @@ public class TopologicalStyleProvider extends AbstractBaseVoltageDiagramStylePro
         if (node.getType() == NodeType.SWITCH && ((SwitchNode) node).isOpen()) {
             return Optional.of(DiagramStyles.DISCONNECTED_STYLE_CLASS);
         }
-        String style = super.getVoltageLevelNodeStyle(voltageLevelInfos, node)
-                .flatMap(baseVoltageLevelStyle -> getNodeTopologicalStyle(baseVoltageLevelStyle, voltageLevelInfos.getId(), node))
+        String style = baseVoltagesConfig.getBaseVoltageName(voltageLevelInfos.getNominalVoltage(), BASE_VOLTAGE_PROFILE)
+                .flatMap(baseVoltageName -> getNodeTopologicalStyle(DiagramStyles.STYLE_PREFIX + baseVoltageName, voltageLevelInfos.getId(), node))
                 .orElse(DiagramStyles.DISCONNECTED_STYLE_CLASS);
         return Optional.of(style);
     }
